added note that one needs to click confirm in first email from listserv
Some folks say they signed up for the listserv but they don't appear in the listserv subscribers list. Following up with some users (thanks Kojun!), we discovered that they need to click "confirm" in the first email they receive from the listserv. Added note to this effect here on the events.html page, and will add it to the listserv instructions page too.