change calendar order, add text about Chicago time
Note to @LocoDelAssembly I'm not sure about the UTC option. I think (will take experiment), I'd need to go back to recreate (or edit) the time zone of the entire calendar -- to use UTC. Then maybe this will work for all 3 calendar options. At the very least, it would get the "public URL" calendar to be in UTC. Our other option is to add javascript for the Public URL.